refactor(executor): Split simd_filter.rs into smaller modules
Split the 1589-line simd_filter.rs file into logical modules for better
maintainability:

- mod.rs (~200 lines code + ~540 lines tests): Main API and dispatch logic
- comparison.rs (395 lines): Numeric comparison operations (i64, i32, f64)
- string.rs (164 lines): String-specific filter operations
- mask.rs (244 lines): Filter mask application to batches
- conversion.rs (83 lines): Value type conversion utilities

All new files are under 400 lines as requested. The public API remains
unchanged - only the internal organization has been improved.

Closes #3052
